+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Speed improvement for `*_gaps()` and `group_by_key()`.
 * `index_by()` allows for grouping index variable; and `group_by()` will throw out an error for grouping index.
+* `index_by(.data, <empty>)` by default groups the index variable rather than previously `ungroup()`.
 * When creating a tsibble, a warning might be issued for mismatch between interval and index representation.
 * Deprecate `as.tsibble()` in favour of `as_tsibble()`.
 * Defunct `fill_na()` in favour of `fill_gaps()`.